Refactored render code slightly.
Separated rendering from interval handling so that first frame can
be rendered as soon as it's loaded. (This will change later because
we don't want to assume all images for all players will load when
the page loads, in case a single page has dozens of animations with
hundreds of frames each.)

Added placeholder stop/start event handlers so the interval won't
run continuously.

Experimented with using separate memory canvases to "pre-render"
images. This doesn't seem to make much difference and needs to be
benchmarked across browsers to see if this matters (I suspect not
since the base Image object is probably as efficient as a canvas).

Disabled some of the more verbose debug output.