Refactored frame position computation.
Previously frame position was computed incrementally, with speed
and position adjusted with each invocation to interval(). This had
two main problems:

  1. Acceleration values are tied to frame rate, and must be scaled
     if the frame rate changes. (This is not a big deal, but it does
     mean the frame rate is not expected to fluctuate.)

  2. Anything that causes the playback to stutter (such as something
     else the browser or host computer might be doing) will actually
     cause frames to be delayed, not dropped.

Most importantly, #1 implies that requestAnimationFrame() cannot be
used because we "know" how long it will be until we next render a
frame. Using rAF is desirable because it automatically scales the
frame rate to whatever the target browser can handle.

To accommodate this expected improvement, the code needs to shift
from knowing its position to instead knowing its time, and computing
its position from its time and the position/speed of the animation at
the start of this reference period. This commit includes this code.

This code contains a bug where time past the "end" of the animation
simply reverses the animation and accelerates continuously back to
the beginning and then crashes.